+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * End screen capture when SDL renderer is destroyed
 * Register all connections to X server (for savestates) instead of just one
 * Fix screen capture of SDL1 surface
+* Autosave and exit save when movie is modified in the input editor
 
 ## [1.3.1] - 2018-08-30
 ### Added